make the box move

+++ GameLib/Vector2.cs
@@ -2,7 +2,7 @@
 {
     public class Vector2
     {
-        public int x, y;
+        public float x, y;
 
         public Vector2() 
         {
@@ -10,7 +10,7 @@
             y = 0;
         }
 
-        public Vector2(int x, int y)
+        public Vector2(float x, float y)
         {
             this.x = x;
             this.y = y;
